+++ ChangeLog.md
@@ -2,6 +2,7 @@
   * Update documentation for material and rarity information
   * Update coriolis export for Beluga Liner
   * Help large star system databases by adding an index
+  * Add 'shared' flag for bounty events
 
 #2.0.9
   * Do not update ship configuration when data is not available from the companion API